Add_song: move quantisation functions to separate file
